added lots of work towards adding SIG support and homepage shuffel
