+++ .gitignore
@@ -67,3 +67,4 @@ dev.db
 .idea
 htdocs/static/
 htdocs/media/
+chipy_org/mediafiles/
