+++ chipy_org/apps/meetings/templates/meetings/meeting.html
@@ -9,13 +9,12 @@
 {% block body %}
 <div itemscope itemtype="http://schema.org/Event">
 
-
     <div class="row-fluid">
       <div class="span12">
 
-
           {% if meeting.meeting_type %}
           <h3 itemprop="name">{{ meeting.meeting_type.name }}</h3>
+          {% if meeting.meeting_type.subgroup %}<p><a href="{% url 'groups' meeting.meeting_type.subgroup.slug %}">{{ meeting.meeting_type.subgroup.name }}</a></p>{% endif %}
           {% endif %}
           <hr />
           {% if meeting.description %}
@@ -50,14 +49,13 @@
           {% endif %}
 
       </div>
-    </div>
 
+  </div>
+  <div class="row-fluid">
 
-
-    <div class="row-fluid">
-
-        <div class="module span12">
+      <div class="module span12">
             <h4>Topics</h4>
+            <hr />
             <ul>
               {% for topic in meeting.topics_list %}
                 <li>
@@ -75,7 +73,8 @@
                 </li>
               {% endfor %}
             </ul>
-        </div><!--/span-->
+       </div>
+
     </div><!-- #row-fluid -->
 
 </div>
